added authentication, middleware and user to listings relationships

+++ routes/web.php
@@ -5,50 +5,61 @@ use App\Http\Controllers\UserController;
 use App\Models\Listing;
 use Illuminate\Support\Facades\Route;
 
-// All Listings
+/*
+|--------------------------------------------------------------------------
+| Web Routes
+|--------------------------------------------------------------------------
+|
+| Here is where you can register web routes for your application. These
+| routes are loaded by the RouteServiceProvider within a group which
+| contains the "web" middleware group. Now create something great!
+|
+*/
 
+// Common Resource Routes:
+// index - Show all listings
+// show - Show single listing
+// create - Show form to create new listing
+// store - Store new listing
+// edit - Show form to edit listing
+// update - Update listing
+// destroy - Delete listing
+
+// All Listings
 Route::get('/', [ListingController::class, 'index']);
 
 // Show Create Form
-
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 
 // Store Listing Data
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 
-Route::post('/listings', [ListingController::class, 'store']);
-
-// Show Edit form
-
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+// Show Edit Form
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 
 // Update Listing
-
-Route::put('/listings/{listing}', [ListingController::class, 'update']);
+Route::put('/listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 
 // Delete Listing
+Route::delete('/listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 
-Route::delete('/listings/{listing}', [ListingController::class, 'destroy']);
-
-// Show Register/Create Form
-
-Route::get('/register', [UserController::class, 'create']);
+// Manage Listings
+Route::get('/listings/manage', [ListingController::class, 'manage'])->middleware('auth');
 
 // Single Listing
-
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
-// Create New User
+// Show Register/Create Form
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 
+// Create New User
 Route::post('/users', [UserController::class, 'store']);
 
 // Log User Out
-
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 
 // Show Login Form
-
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');
 
 // Log In User
-
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
